+++ diesel/src/sqlite/connection/functions.rs
@@ -40,6 +40,24 @@ where
     Ok(())
 }
 
+pub fn register_noargs<RetSqlType, Ret, F>(
+    conn: &RawConnection,
+    fn_name: &str,
+    deterministic: bool,
+    mut f: F,
+) -> QueryResult<()>
+where
+    F: FnMut() -> Ret + std::panic::UnwindSafe + Send + 'static,
+    Ret: ToSql<RetSqlType, Sqlite>,
+    Sqlite: HasSqlType<RetSqlType>,
+{
+    conn.register_sql_function(fn_name, 0, deterministic, move |_, _| {
+        let result = f();
+        process_sql_function_result::<RetSqlType, Ret>(result)
+    })?;
+    Ok(())
+}
+
 pub fn register_aggregate<ArgsSqlType, RetSqlType, Args, Ret, A>(
     conn: &RawConnection,
     fn_name: &str,