+++ diesel_derives/src/sql_function.rs
@@ -302,6 +302,66 @@ pub(crate) fn expand(input: SqlFunctionDecl) -> Result<TokenStream, Diagnostic> 
                 }
             };
         }
+
+        if cfg!(feature = "sqlite")
+            && type_args.is_empty()
+            && arg_name.is_empty()
+            && is_sqlite_type(&return_type)
+        {
+            tokens = quote! {
+                #tokens
+
+                use diesel::sqlite::{Sqlite, SqliteConnection};
+                use diesel::serialize::ToSql;
+
+                #[allow(dead_code)]
+                /// Registers an implementation for this function on the given connection
+                ///
+                /// This function must be called for every `SqliteConnection` before
+                /// this SQL function can be used on SQLite. The implementation must be
+                /// deterministic (returns the same result given the same arguments). If
+                /// the function is nondeterministic, call
+                /// `register_nondeterministic_impl` instead.
+                pub fn register_impl<F, Ret>(
+                    conn: &SqliteConnection,
+                    f: F,
+                ) -> QueryResult<()>
+                where
+                    F: Fn() -> Ret + std::panic::UnwindSafe + Send + 'static,
+                    Ret: ToSql<#return_type, Sqlite>,
+                {
+                    conn.register_noarg_sql_function::<#return_type, _, _>(
+                        #sql_name,
+                        true,
+                        f,
+                    )
+                }
+
+                #[allow(dead_code)]
+                /// Registers an implementation for this function on the given connection
+                ///
+                /// This function must be called for every `SqliteConnection` before
+                /// this SQL function can be used on SQLite.
+                /// `register_nondeterministic_impl` should only be used if your
+                /// function can return different results with the same arguments (e.g.
+                /// `random`). If your function is deterministic, you should call
+                /// `register_impl` instead.
+                pub fn register_nondeterministic_impl<F, Ret>(
+                    conn: &SqliteConnection,
+                    mut f: F,
+                ) -> QueryResult<()>
+                where
+                    F: FnMut() -> Ret + std::panic::UnwindSafe + Send + 'static,
+                    Ret: ToSql<#return_type, Sqlite>,
+                {
+                    conn.register_noarg_sql_function::<#return_type, _, _>(
+                        #sql_name,
+                        false,
+                        f,
+                    )
+                }
+            };
+        }
     }
 
     let args_iter = args.iter();
@@ -391,4 +451,22 @@ impl ToTokens for StrictFnArg {
         self.colon_token.to_tokens(tokens);
         self.name.to_tokens(tokens);
     }
+}
+
+fn is_sqlite_type(ty: &syn::Type) -> bool {
+    [
+        "BigInt",
+        "Binary",
+        "Bool",
+        "Date",
+        "Double",
+        "Float",
+        "Integer",
+        "Numeric",
+        "SmallInt",
+        "Text",
+        "Time",
+        "Timestamp",
+    ]
+    .contains(&quote!(#ty).to_string().as_str())
 }